+++ CHANGELOG.md
@@ -10,8 +10,8 @@
 - Exception can be raised when an unrecognized sorting parameter is received
 
 #### Changed
-- `CreateHydratior` was renamed to `AbstractCreateHydrator`
-- `UpdateHydratior` was renamed to `AbstractUpdateHydrator`
+- `CreateHydrator` was renamed to `AbstractCreateHydrator`
+- `UpdateHydrator` was renamed to `AbstractUpdateHydrator`
 - `AbstractHydrator` can be used for update and create requests too
 - Improved and more extensive documentation